Workflow to publish Helm chart to preprod (#2)
We use `helm package` to create the chart tgz file, and then `helm repo index build` to generate an index.yaml file with the list of available charts. This then gets uploaded to the downloads bucket in a `/helm` directory.

I'm setting the chart version number to `0.0.1-preprod-<timestamp>`, and we'll publish a new version for each merge to main.