Allow chart versions to be overridden
I need to re-run the action to publish v0.11.0 because it failed to publish the `spacelift-workerpool-controller` chart. But unfortunately it's now failing because the other two charts have already been published.

I've added the `--force` flag to allow us to overwrite the published version in this situation since it should only happen if we delete and re-tag a release, in which case we want to re-publish.