feat: add self-hosted chart
This is just a copy and paste of the chart that is currently in our
self-hosted repository. No changes.
I just added a readme with a link to the docs, and I'll update it once
our doc is fully merged with mainstream.
Once the chart is deployed, I'll also update the docs.