feat: simplify the server and mqtt broker service names
We were using the default Helm approach of using the release name as a prefix to the service. But that makes things tricky because we need to pass the name of the MQTT service into an env var.

To make things a bit simpler, I've changed them to use properties from the values files with defaults. That means by default we always use `spacelift-server` and `spacelift-mqtt`, but it's still configurable if required.